// src/models/TipoPropiedad.ts
// noinspection JSUnusedGlobalSymbols
export enum TipoPropiedad{
  CASA = 'Casa',
  DEPARTAMENTO = 'Departamento',
  PH= 'PH',
  LOCAL = 'Local',
  OFICINA = 'Oficina',
  COCHERA = 'Cochera',
  TERRENOS_Y_LOTES = 'Terrenos y Lotes',
  CONSULTORIO = 'Consultorio',
  QUINTA = 'Quinta',
  CHACRA = 'Chacra',
  GALPON = 'Galpón',
  DEPOSITO = 'Depósito',
  CAMPO = 'Campo',
  HOTEL = 'Hotel',
  EDIFICIO = 'Edificio',
  OTRO = 'Otro'
}